Move Accordion to common directory
